requirements.txt
fastapi~=0.115.6
uvicorn~=0.34.0
pydantic~=2.10.0
python-dotenv~=1.1.0
aiofiles~=25.1.0
pytest-cov~=7.0.0
pytest~=9.0.1
pytest-asyncio~=1.3.0
PyYAML~=6.0.2
httpx~=0.28.1
alembic~=1.14.0
graphql-core~=3.2.5
openapi-spec-validator~=0.7.1
sqlalchemy~=2.0.41
asyncpg~=0.30.0
sqlacodegen~=3.1.1

jinja2~=3.1.6
python-multipart~=0.0.20
